fixed ram problems, but dates are corrupt after parsing

+++ Arduino_IDE/SolArduino/numbers.h
@@ -1,5 +1,5 @@
 
 //header file for numbers
 
-extern float angles[];
+extern int angles[];
 extern long dates[];
